test: split test-crypto-dh.js
Split test-crypto-dh.js so that it is less likely to timeout on
less powerful bots.

PR-URL: https://github.com/nodejs/node/pull/40451
Refs: https://github.com/nodejs/reliability/issues/86
